Committing changes on oscar.cbls.lib to migrate to 64bit version

+++ oscar-cbls/src/main/scala/oscar/cbls/lib/invariant/logic/Filter.scala
@@ -35,7 +35,7 @@ import scala.collection.immutable.SortedSet
   * @author renaud.delandtsheer@cetic.be
   * */
 case class Filter(values:Array[IntValue], cond:(Long=>Boolean)=_>0L)
-  extends SetInvariant(values.indices.foldLeft(SortedSet.empty[Long])((acc:SortedSet[Long],indice:Long) => if(cond(values(indice).value)){acc+indice}else acc),
+  extends SetInvariant(values.indices.foldLeft(SortedSet.empty[Long])((acc:SortedSet[Long],indice:Int) => if(cond(values(indice).value)){acc+indice}else acc),
     values.indices.start to values.indices.end)
   with IntNotificationTarget{
